splitting command base into components module
