+++ pydra2app/core/image/tests/test_app.py
@@ -5,13 +5,9 @@ from copy import deepcopy
 from pydra.utils import get_fields
 from frametree.file_system import FileSystem
 from frametree.axes.samples import Samples
-import fileformats.field as ffield
-from fileformats.generic import File
 from pydra2app.core.image import App, P2AImage
-from pydra2app.core.command.base import ContainerCommand, ContainerCommandSink
 from pydra2app.core import PACKAGE_NAME
 from conftest import TestDatasetBlueprint
-import pytest
 
 
 def test_native_python_install(tmp_path: Path) -> None:
@@ -340,45 +336,3 @@ def test_serialization_roundtrip(tmp_path: Path) -> None:
     reloaded_app = App.load(save_path)
     assert app.commands[0] == reloaded_app.commands[0]
     assert app == reloaded_app
-
-
-def test_command():
-
-    cmd = ContainerCommand(
-        **{
-            "task": {
-                "type": "shell",
-                "executable": [
-                    "pydra2app",
-                    "--version<print_version>",
-                ],
-                "inputs": {
-                    "dummy": {
-                        "type": int | None,
-                        "help": "not actually used",
-                        "argstr": None,  # won't be printed to the command line
-                    }
-                },
-            },
-            "operates_on": "samples/sample",
-            "sinks": {"pydra2app_version": "stdout"},
-        }
-    )
-
-    assert len(cmd.sinks) == 1
-    sink = cmd.sinks[0]
-    assert sink.name == "pydra2app_version"
-    assert sink.field == "stdout"
-    assert sink.type is ffield.Text
-    assert len(cmd.parameters) == 3
-    params = {p.name: p for p in cmd.parameters}
-    assert set(params) == {"dummy", "print_version", "append_args"}
-    assert params["dummy"].type == ffield.Integer | None
-    assert params["dummy"].help == "not actually used"
-    assert params["print_version"].type == ffield.Boolean
-    assert params["print_version"].help == ""
-    assert params["append_args"].type == list[ffield.Text | File]
-    assert (
-        params["append_args"].help
-        == "Additional free-form arguments to append to the end of the command."
-    )
